Updated CSS to be easier to read, HTML to use a form, and started adding review info to the console.
For the CSS update, the bug where only some of the bar would allow text on it is fixed, as well as other things that make the page look better. For HTML, I removed the console input as it was optional anyways, and the API that I'm using now does not have that as an input option. For the rest of the project, I'm using a new API that just gets data from IGN as the previous one I was using seems to be broken. While this is unfortunate, I will still be using the information from GiantBomb, so there will still be reviews from two sources. I also started to get the mock-data (which is based on the way the API pulls data from IGN) to print out onto the console of the web page.